requirements.txt
asgiref==3.8.1
Django==4.2
django-ranged-response==0.2.0
django-robots==6.1
django-simple-captcha==0.6.0
pillow==10.4.0
sqlparse==0.5.0
tzdata==2024.1
